05_Quinto_Periodo/doceriAnaDiniz/src/backend/src/test/java/com/br/pucminas/backend/config/SecurityConfigTest.java
package com.br.pucminas.backend.config;

import org.junit.Before;
import org.junit.Test;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.security.crypto.password.PasswordEncoder;

import java.security.SecureRandom;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;
import static org.mockito.Mockito.when;

public class SecurityConfigTest {

    @InjectMocks
    private SecurityConfig securityConfig;

    @Before
    public void init() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testPasswordEncoder() {
        PasswordEncoder passwordEncoder = securityConfig.passwordEncoder();

        assertNotNull(passwordEncoder);

        // You can add more test cases here to check the behavior of the passwordEncoder.
        // For example, you can test password encoding and matching.
    }
}
